api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
h3>=4.0.0
